determine hashmap dtype by N*W*H*D instead of coords.dtype

+++ flex_gemm/kernels/cuda/spconv/neighbor_map.cu
@@ -96,7 +96,7 @@ __global__ void hashmap_lookup_submanifold_conv_neighbour_map_cuda_kernel(
  */
 __global__ void hashmap_lookup_submanifold_conv_neighbour_map_cuda_kernel_64(
     const uint64_t N,
-    const uint64_t M,
+    const uint32_t M,
     const int W,
     const int H,
     const int D,
@@ -108,7 +108,7 @@ __global__ void hashmap_lookup_submanifold_conv_neighbour_map_cuda_kernel_64(
     const int Dh,
     const int Dd,
     const uint64_t* __restrict__  hashmap,
-    const int64_t* __restrict__  coords,
+    const int32_t* __restrict__  coords,
     uint64_t* __restrict__ neighbor
 ) {
     int64_t thread_id = blockIdx.x * blockDim.x + threadIdx.x;
@@ -242,7 +242,7 @@ torch::Tensor hashmap_build_submanifold_conv_neighbour_map_cuda(
             Dh,
             Dd,
             hashmap.data_ptr<uint64_t>(),
-            coords.data_ptr<int64_t>(),
+            coords.data_ptr<int32_t>(),
             neighbor.data_ptr<uint64_t>()
         );
         return neighbor;
